pyproject.toml
[tool.poetry]
name = "job-easy-find"
version = "0.1.0"
description = ""
authors = ["fthh <sample.fthh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
